create new log module for logger
