+++ packages/logger/src/index.ts
@@ -1,5 +1,4 @@
-import { arrayEquals, asArray } from "@solid-primitives/utils"
-import { getOwner, OwnerType, SolidComputation, SolidOwner, SolidSignal } from "@shared/graph"
+import { getOwner, OwnerType, SolidComputation, SolidSignal } from "@shared/graph"
 import {
 	getOwnerType,
 	getOwnerName,
@@ -8,23 +7,19 @@ import {
 	onParentCleanup,
 	getFunctionSources,
 	makeSolidUpdateListener,
+	getName,
 } from "@solid-devtools/debugger"
-
-type UpdateCause = {
-	type: "signal" | "computation"
-	name: string
-	value: unknown
-}
-
-const UNUSED = Symbol("unused")
-
-type ComputationState = {
-	owned: SolidComputation[]
-	prev: unknown | typeof UNUSED
-	value: unknown | typeof UNUSED
-	sources: (SolidComputation | SolidSignal)[]
-	causedBy: UpdateCause[] | null
-}
+import {
+	getComputationCreatedLabel,
+	getComputationRerunLabel,
+	getOwnerDisposedLabel,
+	logComputation,
+	logInitialValue,
+	logObservers,
+	logSignalValueUpdate,
+	UNUSED,
+	UpdateCause,
+} from "./log"
 
 declare module "solid-js/types/reactive/signal" {
 	interface Owner {
@@ -35,21 +30,11 @@ declare module "solid-js/types/reactive/signal" {
 	}
 }
 
-const STYLES = {
-	bold: "font-weight: bold; font-size: 1.1em;",
-	ownerName:
-		"font-weight: bold; font-size: 1.1em; background: rgba(153, 153, 153, 0.3); padding: 0.1em 0.3em; border-radius: 4px;",
-	grayBackground: "background: rgba(153, 153, 153, 0.3); padding: 0 0.2em; border-radius: 4px;",
-	signalUnderline: "text-decoration: orange wavy underline;",
-}
+const dedupeArray = <T>(array: readonly T[]) => Array.from(new Set(array))
 
-const inGray = (text: unknown) => `\x1B[90m${text}\x1B[m`
-const styleTime = (time: number) => `\x1B[90;3m${time} ms\x1B[m`
-
-const dedupeArray = <T>(array: T[]) => Array.from(new Set(array))
-
-const getName = (o: SolidSignal | SolidOwner) =>
-	isComputation(o) ? getOwnerName(o) : o.name ?? "(unnamed)"
+/** Checks if both arrays contain the same values. Order doesn't matter. */
+const arrayRefEquals = <T>(a: readonly T[], b: readonly T[]) =>
+	a === b || (a.length === b.length && a.every(e => b.includes(e)))
 
 const makeTimeMeter = () => {
 	let last = performance.now()
@@ -59,73 +44,6 @@ const makeTimeMeter = () => {
 		last = now
 		return Math.round(diff)
 	}
-}
-
-const logComputationDetails = ({
-	causedBy,
-	owned,
-	sources,
-	prev,
-	value,
-}: Readonly<ComputationState>) => {
-	// Value
-	if (value !== UNUSED) console.log(inGray("Value ="), value)
-	if (prev !== UNUSED) console.log(inGray("Previous ="), prev)
-
-	// Caused By
-	if (causedBy && causedBy.length) {
-		if (causedBy.length === 1)
-			console.log(
-				`%c${inGray("Caused By:")} %c${causedBy[0].name}%c ${inGray("=")}`,
-				"",
-				"owned" in causedBy[0] ? STYLES.grayBackground : STYLES.signalUnderline,
-				"",
-				causedBy[0].value,
-			)
-		else {
-			console.groupCollapsed(inGray("Caused By:"), causedBy.length)
-			causedBy.forEach(cause => {
-				console.log(
-					`%c${cause.name}%c ${inGray("=")}`,
-					cause.type === "computation" ? STYLES.grayBackground : STYLES.signalUnderline,
-					"",
-					cause.value,
-				)
-			})
-			console.groupEnd()
-		}
-	}
-
-	// Sources
-	if (sources.length) {
-		console.groupCollapsed(inGray("Sources:"), sources.length)
-		sources.forEach(source => {
-			console.log(
-				`%c${getName(source)}%c ${inGray("=")}`,
-				isComputation(source) ? STYLES.grayBackground : STYLES.signalUnderline,
-				"",
-				source.value,
-			)
-		})
-		console.groupEnd()
-	} else {
-		console.log(inGray("Sources:"), 0)
-	}
-
-	// Owned
-	if (owned.length) {
-		console.groupCollapsed(inGray("Owned:"), owned.length)
-		owned.forEach(owned => console.log(`%c${owned.name}`, STYLES.grayBackground))
-		console.groupEnd()
-	} else {
-		console.log(inGray("Owned:"), 0)
-	}
-}
-
-const logComputation = (groupLabel: string | string[], state: Readonly<ComputationState>) => {
-	console.groupCollapsed(...asArray(groupLabel))
-	logComputationDetails(state)
-	console.groupEnd()
 }
 
 export function debugComputation() {
@@ -177,16 +95,13 @@ export function debugComputation() {
 			removeValueObserver()
 			const sources = owner.sources ? dedupeArray(owner.sources) : []
 
-			logComputation(
-				[`%c${typeName} %c${name}%c created  ${styleTime(timeElapsed)}`, "", STYLES.ownerName, ""],
-				{
-					owned: owner.owned ?? [],
-					sources,
-					prev: UNUSED,
-					value: usesValue ? value : UNUSED,
-					causedBy: null,
-				},
-			)
+			logComputation(getComputationCreatedLabel(typeName, name, timeElapsed), {
+				owned: owner.owned ?? [],
+				sources,
+				prev: UNUSED,
+				value: usesValue ? value : UNUSED,
+				causedBy: null,
+			})
 			observeSources(sources)
 		},
 		SYMBOL,
@@ -203,7 +118,7 @@ export function debugComputation() {
 		const elapsedTime = time()
 
 		const sources = owner.sources ? dedupeArray(owner.sources) : []
-		logComputation([`%c${name}%c re-executed  ${styleTime(elapsedTime)}`, STYLES.ownerName, ""], {
+		logComputation(getComputationRerunLabel(name, elapsedTime), {
 			owned: owner.owned ?? [],
 			sources,
 			prev: usesPrev ? prev : UNUSED,
@@ -220,7 +135,7 @@ export function debugComputation() {
 	onParentCleanup(
 		owner,
 		() => {
-			console.log(`%c${name}%c disposed`, STYLES.ownerName, "")
+			console.log(getOwnerDisposedLabel(name))
 			updateListeners.forEach(unsub => unsub())
 			updateListeners.length = 0
 			signalUpdates.length = 0
@@ -232,79 +147,10 @@ export function debugComputation() {
 	const time = makeTimeMeter()
 }
 
-function logCausedUpdates(observers: SolidComputation[]): void {
-	if (!observers.length) return
-	console.groupCollapsed(inGray("Caused Updates:"), observers.length)
-	const toLog: [type: string, name: string][] = []
-	let typeLength = 0
-	observers.forEach(observer => {
-		const type = OwnerType[getOwnerType(observer)]
-		typeLength = Math.max(type.length, typeLength)
-		toLog.push([type, getName(observer)])
-	})
-	toLog.forEach(([type, name]) => {
-		console.log(`${inGray(type.padEnd(typeLength))} %c${name}`, STYLES.grayBackground)
-	})
-	console.groupEnd()
-}
-
-function logObservers(
-	signalName: string,
-	observers: SolidComputation[],
-	prevObservers: SolidComputation[],
+export function debugSignal(
+	getter: () => unknown,
+	{ trackObservers = true }: { trackObservers?: boolean } = {},
 ): void {
-	const label = [
-		`%c${signalName}%c observers changed:`,
-		`${STYLES.bold} ${STYLES.signalUnderline}`,
-		"",
-		observers.length,
-	]
-	if (!observers.length && !prevObservers.length) return console.log(...label)
-	console.groupCollapsed(...label)
-	const toLog: [type: string, name: string, mark: "added" | "removed" | null][] = []
-	let typeLength = 0
-	prevObservers.forEach(observer => {
-		const type = OwnerType[getOwnerType(observer)]
-		typeLength = Math.max(type.length, typeLength)
-		toLog.push([type, getName(observer), observers.includes(observer) ? null : "removed"])
-	})
-	observers.forEach(observer => {
-		if (prevObservers.includes(observer)) return
-		const type = OwnerType[getOwnerType(observer)]
-		typeLength = Math.max(type.length, typeLength)
-		toLog.push([type, getName(observer), "added"])
-	})
-	toLog.forEach(([type, name, mark]) => {
-		const formattedType = inGray(type.padEnd(typeLength))
-		switch (mark) {
-			case "added":
-				console.log(
-					`${formattedType} %c${name}%c  new`,
-					STYLES.grayBackground,
-					"color: orange; font-style: italic",
-				)
-				break
-
-			case "removed":
-				console.log(
-					`${formattedType} %c${name}`,
-					"background: rgba(153, 153, 153, 0.15); padding: 0 0.2em; border-radius: 4px; text-decoration: line-through; color: #888",
-				)
-				break
-
-			default:
-				console.log(`${formattedType} %c${name}`, STYLES.grayBackground)
-				break
-		}
-	})
-	console.groupEnd()
-}
-
-function logPrevValue(prev: unknown): void {
-	console.log(inGray("Previous:"), prev)
-}
-
-export function debugSignal(getter: () => unknown): void {
 	const sources = getFunctionSources(getter)
 	if (sources.length === 0) return console.warn("No signal was passed to debugSignal")
 	else if (sources.length > 1) return console.warn("More then one signal was passed to debugSignal")
@@ -320,13 +166,7 @@ export function debugSignal(getter: () => unknown): void {
 	const SYMBOL = Symbol(name)
 
 	// Initial
-	console.log(
-		`%c${type} %c${name}%c initial value ${inGray("=")}`,
-		"",
-		`${STYLES.bold} ${STYLES.signalUnderline}`,
-		"",
-		signal.value,
-	)
+	logInitialValue(type, name, signal.value)
 
 	let actualObservers: SolidComputation[]
 	let prevObservers: SolidComputation[] = []
@@ -340,35 +180,33 @@ export function debugSignal(getter: () => unknown): void {
 	// Value Update
 	observeValueUpdate(
 		signal,
-		(value, prev) => {
-			console.groupCollapsed(
-				`%c${name}%c updated ${inGray("=")}`,
-				`${STYLES.bold} ${STYLES.signalUnderline}`,
-				"",
+		(value, prev) =>
+			logSignalValueUpdate(
+				name,
 				value,
-			)
-			logPrevValue(prev)
-			logCausedUpdates(prevObservers)
-			console.groupEnd()
-		},
+				prev,
+				trackObservers ? prevObservers : dedupeArray(signal.observers!),
+			),
 		SYMBOL,
 	)
 
-	// Observers Change
-	function logObserversChange() {
-		const observers = dedupeArray(actualObservers)
-		if (arrayEquals(observers, prevObservers)) return
-		logObservers(name, observers, prevObservers)
-		prevObservers = [...observers]
-		actualPrevObservers = [...actualObservers]
-	}
-
-	// Listen to Solid's _$afterUpdate hook to check if observers changed
-	makeSolidUpdateListener(() => {
-		actualObservers = signal.observers!
-		if (actualObservers.length !== actualPrevObservers.length) return logObserversChange()
-		for (let i = actualObservers.length; i >= 0; i--) {
-			if (actualObservers[i] !== actualPrevObservers[i]) return logObserversChange()
+	if (trackObservers) {
+		// Observers Change
+		function logObserversChange() {
+			const observers = dedupeArray(actualObservers)
+			if (arrayRefEquals(observers, prevObservers)) return
+			logObservers(name, observers, prevObservers)
+			prevObservers = [...observers]
+			actualPrevObservers = [...actualObservers]
 		}
-	})
+
+		// Listen to Solid's _$afterUpdate hook to check if observers changed
+		makeSolidUpdateListener(() => {
+			actualObservers = signal.observers!
+			if (actualObservers.length !== actualPrevObservers.length) return logObserversChange()
+			for (let i = actualObservers.length; i >= 0; i--) {
+				if (actualObservers[i] !== actualPrevObservers[i]) return logObserversChange()
+			}
+		})
+	}
 }
